HP/UX build patch
From <20211002103740.791d7d24@pc09> (private E-Mail) from "H.Merijn
Brand" <linux@tux.freedom.nl> on Sat, Oct 2, 2021. His reply to
questions I had:

    On Wed, 6 Oct 2021 14:39:32 +0200, Ævar Arnfjörð Bjarmason
    <avarab@gmail.com> wrote:

    > Thanks, I managed to get it building with that!
    >
    > Can I add your Signed-off-by for munging/upstreaming these patches, it

    Yes, you can

    > just certifies that you're the copyright holder of them (required for
    > the git project), will make it clear that any testing/mistakes etc.
    > are mine. Will CC you on them if/when that's ready. Of course not all
    > of it's upstream-able, will need to add the relevant things to
    > config.mak.uname etc.
    >
    > Are the /pro/gnu /pro etc. paths something that's just on your system,
    > or typical/common enough of HP/UX to put it into config that'll be
    > used by others on that uname -s?

    No, /pro is Procura-specific (and thus all of my systems)
    It was just convenient enough to abuse it for gnu stuff too, as it
    thusly does not interfere with HP-UX releases

    HP-UX is with a dash, not with a slash

    > If so, are the paths to the GNU tools similarly "standard"? Hopefully

    No, they are not, not even when you happen to get them from the HP
    porting center. They move from /opt to /usr/local back-and-forth on a
    very inconsistent and unpredictable way

    > it can be built/tested mostly with the native toolchain, but the tests
    > e.g. rely on "diff -u" by default for comparison on failure, so
    > pointing just that diff usage to the GNU one would be handy.

    I've seen (and used) a 'g' prefix being rather common on HP-UX and
    AIX, so if you can find gdiff before you can find diff, you can be
    reasonably safe it is the GNU version, but don't pin me on that.

    If I want to build anything serious on AIX or HP-UX, the first thing
    I (re)build is GNU make

    Next to 'diff', note that 'cmp' is really ugly on older versions of AIX
    and HP-UX and ultimately incompatible

+++ compat/regex/regexec.c
@@ -688,7 +688,7 @@ re_search_internal (const regex_t *preg,
   if (nmatch > 1 || dfa->has_mb_node)
     {
       /* Avoid overflow.  */
-      if (BE (SIZE_MAX / sizeof (re_dfastate_t *) <= mctx.input.bufs_len, 0))
+      if (BE ((unsigned long long)SIZE_MAX / sizeof (re_dfastate_t *) <= mctx.input.bufs_len, 0))
 	{
 	  err = REG_ESPACE;
 	  goto free_return;
@@ -952,7 +952,7 @@ prune_impossible_nodes (re_match_context_t *mctx)
   halt_node = mctx->last_node;
 
   /* Avoid overflow.  */
-  if (BE (SIZE_MAX / sizeof (re_dfastate_t *) <= match_last, 0))
+  if (BE ((unsigned long long)SIZE_MAX / sizeof (re_dfastate_t *) <= match_last, 0))
     return REG_ESPACE;
 
   sifted_states = re_malloc (re_dfastate_t *, match_last + 1);
@@ -3372,7 +3372,7 @@ build_trtable (const re_dfa_t *dfa, re_dfastate_t *state)
     goto out_free;
 
   /* Avoid arithmetic overflow in size calculation.  */
-  if (BE ((((SIZE_MAX - (sizeof (re_node_set) + sizeof (bitset_t)) * SBC_MAX)
+  if (BE (((((unsigned long long)SIZE_MAX - (sizeof (re_node_set) + sizeof (bitset_t)) * SBC_MAX)
 	    / (3 * sizeof (re_dfastate_t *)))
 	   < ndests),
 	  0))
